requirements.txt
flask
matplotlib
google-generativeai
beautifulsoup4
requests
